requirements.txt
streamlit
openai
langchain
python-dotenv
google-generativeai
PyPDF2
Chromadb